pyproject.toml
[tool.poetry]
name = "goalforge-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "1.13.1"
annotated-types = "0.6.0"
anyio = "3.7.1"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
bcrypt = "4.0.1"
certifi = "2023.11.17"
cffi = "1.16.0"
click = "8.1.7"
cryptography = "41.0.5"
databases = "0.9.0"
dnspython = "2.4.2"
ecdsa = "0.18.0"
email-validator = "2.1.0.post1"
exceptiongroup = "1.2.0"
fastapi = "0.104.1"
greenlet = "3.0.2"
h11 = "0.14.0"
httpcore = "1.0.2"
httpx = "0.25.2"
idna = "3.4"
iniconfig = "2.0.0"
Mako = "1.3.0"
MarkupSafe = "2.1.3"
nodeenv = "1.8.0"
packaging = "23.2"
passlib = "1.7.4"
pluggy = "1.3.0"
psycopg2-binary = "2.9.9"
pyasn1 = "0.5.1"
pycparser = "2.21"
pydantic = "1.10.13"
pydantic-core = "2.14.5"
python-dotenv = "1.0.0"
python-jose = "3.3.0"
python-multipart = "0.0.6"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
SQLAlchemy = "2.0.29"
sqlmodel = "0.0.16"
starlette = "0.27.0"
tomli = "2.0.1"
typing-extensions = "4.8.0"
uvicorn = "0.24.0.post1"

[tool.poetry.dev-dependencies]
pyright = "1.1.355"
pytest = "7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
